PTFM-18860 Update "dx build" and "dx get" to manipulate multi-region apps
Summary:
This commit changes "dx build" and "dx get" to properly construct and
de-construct multi-region apps. When building an app, "dx build" will inspect
dxapp.json (the "requestedRegionalOptions" key) for instructions on how to
build the possibly multi-region app. When getting an app, "dx get" will
reconstruct "requestedRegionalOptions" key in dxapp.json based on the enabled
regions of the app. No changes are made with respect to building or getting
applets.

Test Plan:
- bin/run_integration_tests.sh --client_tests_only --tests \
  test_dxclient.TestDXBuildApp
- bin/run_integration_tests.sh --client_tests_only --tests \
  test_dxclient.TestDXGetExecutables

Reviewers: psung

Differential Revision: https://dnanexus.phacility.com/D1177